+++ crudUsuario/excluiruser.php
@@ -6,11 +6,7 @@ include_once "../conexao.php";
 $conexao = conectar();
 
 $id = $_GET['id_usuario'];
-$pasta = "../imagens";
-
-$selectImg = "SELECT imagem FROM eventos";
-$result0 = executarSQL($conexao, $selectImg);
-$img = mysqli_fetch_assoc($result0);
+$pasta = "../imagens/";
 
 $select = "SELECT * FROM usuarios WHERE id_usuario='$id'";
 $result = executarSQL($conexao, $select);
@@ -20,46 +16,77 @@ if ($dados['tipo_usuario'] == 2) {
 
     $selectCarrinho = "SELECT id_carrinho FROM carrinhos WHERE id_usuario ='$id'";
     $execSelect = executarSQL($conexao, $selectCarrinho);
-    $resultSelCart = mysqli_fetch_assoc($execSelect);
-    $idCart = $resultSelCart['id_carrinho'];
 
-    $delete_cic = "DELETE FROM carrinho_ingressos_cadastrados WHERE id_carrinho='$idCart'";
+    $idsCarrinhos = [];
+    while ($row = mysqli_fetch_assoc($execSelect)) {
+        $idsCarrinhos[] = $row['id_carrinho'];
+    }
 
-    $delete_inc = "DELETE FROM carrinhos WHERE id_usuario='$id'";
-    $result_DelInc = executarSQL($conexao, $delete_inc);
+    if (!empty($idsCarrinhos)) {
+        $idsCarrinhosStr = implode(",", $idsCarrinhos);
+
+        $delete_cic = "DELETE FROM carrinho_ingressos_cadastrados WHERE id_carrinho IN ($idsCarrinhosStr)";
+        executarSQL($conexao, $delete_cic);
+
+        $delete_inc = "DELETE FROM carrinhos WHERE id_usuario='$id'";
+        executarSQL($conexao, $delete_inc);
+    }
 
     $delete_user = "DELETE FROM usuarios WHERE id_usuario='$id'";
     $result_DelUser = executarSQL($conexao, $delete_user);
 
-    if ($result_DelInc && $result_DelUser) {
+    if ($result_DelUser) {
         $_SESSION['mensagem'][0] = "Usuário excluido com sucesso!";
         $_SESSION['mensagem'][1] = "#558b2f light-green darken-3";
     } else {
         $_SESSION['mensagem'][0] = "Não foi possível excluir o usuário!";
         $_SESSION['mensagem'][1] = "#c62828 red darken-3";
     }
-
 } elseif ($dados['tipo_usuario'] == 3) {
 
-    $delete_incads = "DELETE FROM ingressos_cadastrados WHERE id_usuario='$id'";
-    executarSQL($conexao, $delete_incads);
+    $selectCIC = "SELECT * FROM ingressos_cadastrados ica
+    INNER JOIN eventos e 
+    ON ica.id_evento = e.id_evento WHERE e.id_usuario='$id'";
+    $execSel = executarSQL($conexao, $selectCIC);
 
-    unlink($pasta . $img['imagem']);
+    $idsIngressos = [];
+    while ($result = mysqli_fetch_assoc($execSel)) {
+        $idsIngressos[] = $result['id_ingresso'];
+    }
+
+    if (!empty($idsIngressos)) {
+        $idsIngressosStr = implode(",", $idsIngressos);
+
+        $deleteCIC = "DELETE FROM carrinho_ingressos_cadastrados WHERE id_ingresso IN ($idsIngressosStr)";
+        executarSQL($conexao, $deleteCIC);
+
+        $delete_incads = "DELETE FROM ingressos_cadastrados WHERE id_ingresso IN ($idsIngressosStr)";
+        executarSQL($conexao, $delete_incads);
+    }
+
+    $selectEventos = "SELECT imagem FROM eventos WHERE id_usuario='$id'";
+    $execEventos = executarSQL($conexao, $selectEventos);
+
+    while ($img = mysqli_fetch_assoc($execEventos)) {
+        $caminhoImagem = $pasta . $img['imagem'];
+        if (file_exists($caminhoImagem)) {
+            unlink($caminhoImagem);
+        }
+    }
 
     $delete_eve = "DELETE FROM eventos WHERE id_usuario='$id'";
     executarSQL($conexao, $delete_eve);
 
     $delete_user = "DELETE FROM usuarios WHERE id_usuario='$id'";
-    executarSQL($conexao, $delete_user);
+    $resultDelUser = executarSQL($conexao, $delete_user);
 
-    if ($result_DelInc && $result_DelUser) {
+    if ($resultDelUser) {
         $_SESSION['mensagem'][0] = "Usuário excluido com sucesso!";
         $_SESSION['mensagem'][1] = "#558b2f light-green darken-3";
     } else {
         $_SESSION['mensagem'][0] = "Não foi possível excluir o usuario!";
         $_SESSION['mensagem'][1] = "#c62828 red darken-3";
     }
-
 }
 
 header('location: ../telasAdmin/listarUsers.php');
